pythonserver/requirements.txt
flask
flask_cors
mysql-connector-python
python-dotenv
Flask-Session